feat(working): Cleaned and sorted the CSS in "LandingPage.css". Added shadow management in "App.css". Completed the "return to top" button using JSX and CSS. Optimized "App.jsx" by using "const [isTop, setIsTop] = useContext(false)" and moving the "useEffect" to "LandingPage.jsx".
